[fix] Auth User 순환 참조 제거 - #267
